requirements.txt
flask==3.1.0
flask-sqlalchemy==3.1.1
requests==2.32.3
gunicorn==23.0.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
